fix: change package name from ninja_squad to ninjasquad
fix #13
